requirements.txt
ollama==0.3.2
python-telegram-bot==21.4
web3==6.14.0
python-dotenv==1.0.0